# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TermiPay Admin Console API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AdminUser {
  username: string;
  name: string;
}

export interface User {
  id: number;
  cardUid: string;
  fullName: string;
  contactNumber: string;
  balance: number;
  gcashLoadedTotal: number;
  status: string;
  createdAt: string;
}

export interface CreateUserBody {
  cardUid: string;
  fullName: string;
  contactNumber: string;
  initialBalance: number;
}

export interface UpdateUserBody {
  fullName?: string;
  contactNumber?: string;
  balance?: number;
  status?: string;
}

export interface Transaction {
  id: number;
  timestamp: string;
  cardUid: string;
  fullName: string;
  type: string;
  amount: number;
  status: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  Fare: "Fare",
  "Top-up": "Top-up",
} as const;

export type CreateTransactionBodyStatus =
  (typeof CreateTransactionBodyStatus)[keyof typeof CreateTransactionBodyStatus];

export const CreateTransactionBodyStatus = {
  Success: "Success",
  Failed: "Failed",
  Pending: "Pending",
} as const;

export interface CreateTransactionBody {
  cardUid: string;
  type: CreateTransactionBodyType;
  amount: number;
  status: CreateTransactionBodyStatus;
}

export interface UpdateTransactionBody {
  type?: string;
  amount?: number;
  status?: string;
}

export interface FareRoute {
  id: number;
  origin: string;
  destination: string;
  fareAmount: number;
  isActive: boolean;
}

export interface CreateRouteBody {
  origin: string;
  destination: string;
  fareAmount: number;
}

export interface UpdateRouteBody {
  origin?: string;
  destination?: string;
  fareAmount?: number;
  isActive?: boolean;
}

export interface DashboardStats {
  totalRevenueToday: number;
  totalTapsToday: number;
  registeredCards: number;
  activeRoutes: number;
}

export interface RevenueTrendItem {
  date: string;
  revenue: number;
}

export interface ReportSummary {
  totalRevenue7Days: number;
  averageDailyRevenue: number;
  totalTaps7Days: number;
  dailyBreakdown: RevenueTrendItem[];
}

export type ListUsersParams = {
  search?: string;
};

export type ListTransactionsParams = {
  search?: string;
  type?: ListTransactionsType;
  status?: ListTransactionsStatus;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  Fare: "Fare",
  "Top-up": "Top-up",
} as const;

export type ListTransactionsStatus =
  (typeof ListTransactionsStatus)[keyof typeof ListTransactionsStatus];

export const ListTransactionsStatus = {
  Success: "Success",
  Failed: "Failed",
  Pending: "Pending",
} as const;