# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TermiPay Admin Console API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get current admin user
 */
export const GetMeResponse = zod.object({
  username: zod.string(),
  name: zod.string(),
});

/**
 * @summary List all registered cardholders
 */
export const ListUsersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
});

export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  cardUid: zod.string(),
  fullName: zod.string(),
  contactNumber: zod.string(),
  type: zod.string(), // Added
  balance: zod.number(),
  gcashLoadedTotal: zod.number(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Register a new RFID card
 */
export const CreateUserBody = zod.object({
  cardUid: zod.string(),
  fullName: zod.string(),
  contactNumber: zod.string(),
  initialBalance: zod.number(),
  type: zod.string().optional(), // Added
});

/**
 * @summary Get a user by ID
 */
export const GetUserParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  cardUid: zod.string(),
  fullName: zod.string(),
  contactNumber: zod.string(),
  type: zod.string(), // Added
  balance: zod.number(),
  gcashLoadedTotal: zod.number(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a user
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  fullName: zod.string().optional(),
  contactNumber: zod.string().optional(),
  balance: zod.number().optional(),
  status: zod.string().optional(),
  type: zod.string().optional(), // Added
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  cardUid: zod.string(),
  fullName: zod.string(),
  contactNumber: zod.string(),
  type: zod.string(), // Added
  balance: zod.number(),
  gcashLoadedTotal: zod.number(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a user
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Recently registered users
 */
export const ListRecentUsersResponseItem = zod.object({
  id: zod.number(),
  cardUid: zod.string(),
  fullName: zod.string(),
  contactNumber: zod.string(),
  type: zod.string(), // Added
  balance: zod.number(),
  gcashLoadedTotal: zod.number(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListRecentUsersResponse = zod.array(ListRecentUsersResponseItem);

/**
 * @summary List transactions with filters
 */
export const ListTransactionsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  type: zod.enum(["Fare", "Top-up"]).optional(),
  status: zod.enum(["Success", "Failed", "Pending"]).optional(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  timestamp: zod.coerce.date(),
  cardUid: zod.string(),
  fullName: zod.string(),
  type: zod.string(),
  amount: zod.number(),
  status: zod.string(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a transaction
 */
export const CreateTransactionBody = zod.object({
  cardUid: zod.string(),
  type: zod.enum(["Fare", "Top-up"]),
  amount: zod.number(),
  status: zod.enum(["Success", "Failed", "Pending"]),
});

/**
 * @summary Update a transaction
 */
export const UpdateTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTransactionBody = zod.object({
  type: zod.string().optional(),
  amount: zod.number().optional(),
  status: zod.string().optional(),
});

export const UpdateTransactionResponse = zod.object({
  id: zod.number(),
  timestamp: zod.coerce.date(),
  cardUid: zod.string(),
  fullName: zod.string(),
  type: zod.string(),
  amount: zod.number(),
  status: zod.string(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all fare routes
 */
export const ListRoutesResponseItem = zod.object({
  id: zod.number(),
  origin: zod.string(),
  destination: zod.string(),
  fareAmount: zod.number(),
  isActive: zod.boolean(),
});
export const ListRoutesResponse = zod.array(ListRoutesResponseItem);

/**
 * @summary Add a new fare route
 */
export const CreateRouteBody = zod.object({
  origin: zod.string(),
  destination: zod.string(),
  fareAmount: zod.number(),
});

/**
 * @summary Update a fare route
 */
export const UpdateRouteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateRouteBody = zod.object({
  origin: zod.string().optional(),
  destination: zod.string().optional(),
  fareAmount: zod.number().optional(),
  isActive: zod.boolean().optional(),
});

export const UpdateRouteResponse = zod.object({
  id: zod.number(),
  origin: zod.string(),
  destination: zod.string(),
  fareAmount: zod.number(),
  isActive: zod.boolean(),
});

/**
 * @summary Delete a fare route
 */
export const DeleteRouteParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle active status of a route
 */
export const ToggleRouteParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleRouteResponse = zod.object({
  id: zod.number(),
  origin: zod.string(),
  destination: zod.string(),
  fareAmount: zod.number(),
  isActive: zod.boolean(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardStatsResponse = zod.object({
  totalRevenueToday: zod.number(),
  totalTapsToday: zod.number(),
  registeredCards: zod.number(),
  activeRoutes: zod.number(),
});

/**
 * @summary Get 7-day revenue trend
 */
export const GetRevenueTrendResponseItem = zod.object({
  date: zod.string(),
  revenue: zod.number(),
});
export const GetRevenueTrendResponse = zod.array(GetRevenueTrendResponseItem);

/**
 * @summary Get report summary with 7-day analytics
 */
export const GetReportSummaryResponse = zod.object({
  totalRevenue7Days: zod.number(),
  averageDailyRevenue: zod.number(),
  totalTaps7Days: zod.number(),
  dailyBreakdown: zod.array(
    zod.object({
      date: zod.string(),
      revenue: zod.number(),
    }),
  ),
});