# lib/api-zod/src/generated/types/createRouteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TermiPay Admin Console API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRouteBody {
  origin: string;
  destination: string;
  fareAmount: number;
}